Support word document creation.
Modified main article so that it is just a combination
of each section with nice formatting around it. Then added
a second "simple_doc" that combines the sections but turns
off the figures so that it looks nice when converted to
a .docx file. The only missing feature is to copy the figures
from the main article into the docx file. I'll do it manually for
now.